% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-alley.R
\docType{data}
\name{alley.temperature}
\alias{alley.temperature}
\alias{alley.accumulation}
\title{GISP2 Ice Core Temperature and Accumulation Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1632 rows and 2 columns.}
\source{
Supplementary material of Alley, Richard B. 2000. “The Younger Dryas
 Cold Interval as Viewed from Central Greenland.” Quaternary Science Reviews
 19 (1–5): 213–26. doi:10.1016/S0277-3791(99)00062-1.
}
\usage{
data(alley.temperature)
data(alley.accumulation)
}
\description{
Temperature interpretation based on stable isotope analysis, and 
 ice accumulation data, from the GISP2 ice core, central Greenland. Data are 
 smoothed from original measurements published by Cuffey and Clow (1997), as 
 presented in Figure 1 of Alley (2000).

LAST UPDATE: 3/2004 (Original Receipt by WDC Paleo)

NOTE: PLEASE CITE ORIGINAL REFERENCE WHEN USING THIS DATA!!!!!

CONTRIBUTOR:  Richard Alley, Pennsylvania State University.
IGBP PAGES/WDCA CONTRIBUTION SERIES NUMBER: 2004-013

GEOGRAPHIC REGION: Greenland
PERIOD OF RECORD: 49 KYrBP - present
}
\section{Data}{


alley.temperature
Temperature in central Greenland  
Column 1: Age (thousand years before present)
Column 2: Temperature in central Greenland (degrees C) 

alley.accumulation
Accumulation rate in central Greenland
Column 1: Age (thousand years before present) 
Column 2: Accumulation rate (m. ice/year)
}

\section{SUGGESTED DATA CITATION}{
 Alley, R.B..  2004.
GISP2 Ice Core Temperature and Accumulation Data. 
IGBP PAGES/World Data Center for Paleoclimatology 
Data Contribution Series #2004-013. 
NOAA/NGDC Paleoclimatology Program, Boulder CO, USA.
}

\section{ORIGINAL REFERENCE}{
 Alley, R.B. 2000. 
The Younger Dryas cold interval as viewed from central Greenland. 
Quaternary Science Reviews 19:213-226.
}

\section{ADDITIONAL REFERENCE}{

Cuffey, K.M., and G.D. Clow. 1997. Temperature, accumulation, and ice sheet
elevation in central Greenland through the last deglacial transition.
Journal of Geophysical Research 102:26383-26396.
}

\section{Abstract}{
 Greenland ice-core records provide an exceptionally clear picture of 
many aspects of abrupt climate changes, and particularly of those 
associated with the Younger Dryas event, as reviewed here. 
Well-preserved annual layers can be counted confidently, with only 1% 
errors for the age of the end of the Younger Dryas 11,500 years before 
present. Ice-flow corrections allow reconstruction of snow accumulation 
rates over tens of thousands of years with little additional uncertainty. 
Glaciochemical and particulate data record atmospheric-loading changes 
with little uncertainty introduced by changes in snow accumulation. 
Confident paleothermometry is provided by site-specific calibrations 
using ice-isotopic ratios, borehole temperatures, and gas-isotopic ratios. 
Near-simultaneous changes in ice-core paleoclimatic indicators of local, 
regional, and more-widespread climate conditions demonstrate that much 
of the Earth experienced abrupt climate changes synchronous with 
Greenland within thirty years or less. Post-Younger Dryas changes 
have not duplicated the size, extent and rapidity of these 
paleoclimatic changes.
}

\keyword{datasets}
