% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{climproxyrecords}
\alias{climproxyrecords}
\alias{climproxyrecords-package}
\title{climproxyrecords: A package containing climate proxy data.}
\description{
The climproxyrecords package supplies proxy climate record data from
published compilations. So far it includes data from two compilations:
Marcott et al (2013),  Shakun et al (2012), and the GISP2 Ice Core 
Temperature and Accumulation Data from Alley (2000).
}
\details{
For each compilation the package contains three dataframes: \code{proxies},
\code{dating}, and \code{metadata} (e.g. \code{marcott.proxies}). These
contain the proxy data, associated carbon dating information, and metadata
respectively.

Please cite both the compilaton papers (Marcott et al 2013, Shakun et al
2012), and the original sources from which the data were extracted. A full
list of sources can be found in the dataframes \code{compilation.metadata}.

For the GISP2 data please cite Alley, R.B. 2000. 
The Younger Dryas cold interval as viewed from central Greenland. 
Quaternary Science Reviews 19:213-226. and see \code{help(alley.temperature)}
}
