% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{herbert.ODP846.temperature}
\alias{herbert.ODP846.temperature}
\title{Alkenone based temperature reconstruction from core ODP846 (-3.095, -90.817) by Herbert et al (2010)}
\format{
A tbl (data.frame) with 3 columns and 2250 rows containing the 
 following variables: \tabular{lll}{ \bold{Variable.name} \tab \bold{Units}  \tab \bold{Description} 
 \cr Depth                  \tab m    \tab Depth in metres (RMCD)                                       
 \cr Age                   \tab kyr BP \tab Age in kyr BP, Present = 1950 AD
 \cr Temperature           \tab deg C \tab Uk'37 SST, °C
 }
}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/contributions_by_author/herbert2010/herbert2010.txt}
}
\usage{
data(herbert.ODP846.temperature)
}
\description{

}
\section{Reference}{
 
Herbert, T.D., L.C. Peterson, K.T. Lawrence, and Z. Liu. 2010. 
Tropical Ocean Temperatures over the Past 3.5 Myr.
Science, Vol. 328, no. 5985, pp. 1530-1534, 18 June 2010. 
DOI: 10.1126/science.1185435
}

\section{Corrections}{
 None so far required
}

\keyword{datasets}
