% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{marcott2013}
\alias{marcott2013}
\alias{marcott2013-package}
\title{marcott2013: A package containing the climate proxy data from Marcott et al
(2013).}
\description{
The marcott2013 package contains three dataframes: \code{proxies},
\code{dating}, and \code{metadata.} These contain the proxy data, associated carbon dating
information, and metadata respectively.
}
\details{
Please cite both Marcott et al (2013), and the original sources from which
the data were extracted. A full list of sources can be found in the
dataframe \code{metadata}
}

