% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{naafs.U1313.temperature}
\alias{naafs.U1313.temperature}
\title{Alkenone based temperature reconstruction from core U1313 by Naafs et al (2012)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2455 rows and 10 columns.}
\source{
\url{https://doi.org/10.1594/PANGAEA.757946}
}
\usage{
data(naafs.U1313.temperature)
}
\description{

}
\section{Reference}{

}

\section{Corrections}{
 None so far required
}

\keyword{datasets}
