% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{shakun.proxies}
\alias{shakun.proxies}
\title{Climate proxy data from Shakun et al (2012)}
\format{A tbl (data.frame) with 14 columns and 16879 rows containing the 
 following variables: \tabular{lll}{ \bold{Variable.name} \tab \bold{Units}  \tab \bold{Description} 
 \cr Core                  \tab NA    \tab Core ID                                            
 \cr Number                \tab NA    \tab ID number assigned in publication                                           
 \cr ID                    \tab NA    \tab A unique ID formed by combining Core and Proxy.type
 \cr Proxy.type            \tab NA    \tab Proxy type                                         
 \cr Proxy.value           \tab NA    \tab Proxy value                           
 \cr Published.temperature \tab deg C \tab NA                                
 \cr Temperature           \tab deg C \tab NA        
 \cr Age                   \tab yr BP \tab Best available age in yr BP, Present = 1950 AD, Marine04.age when present, otherwise Published.age   
 \cr Published.age         \tab yr BP \tab Age in yr BP, Present = 1950 AD   
 \cr Age.model.error       \tab yr    \tab Estimated error in age model, in years per 1 SD    
 \cr Marine04.age          \tab yr BP \tab Age in yr BP, Present = 1950 AD 
 \cr Marine09.age          \tab yr BP \tab Age in yr BP, Present = 1950 AD       
 \cr Proxy.depth.m         \tab m     \tab Proxy depth   
 \cr Proxy.depth.cm        \tab cm    \tab Proxy depth 
 }}
\source{
\url{http://www.nature.com/nature/journal/v484/n7392/full/nature10915.html}
}
\usage{
data(shakun.proxies)
}
\description{
The climate proxy data supplied as supplementary online material 
 with the publication, Shakun et al (2012), was read into R from the .xls 
 file and reformatted into a handy format for use in R.
 
 Please cite both Shakun et al (2012), and the original sources from which 
 the data were extracted. A full list of sources can be found in the
 dataframe \code{shakun.metadata}
}
\section{Reference}{
 Shakun, Jeremy D., Peter U. Clark, Feng He, Shaun A.
 Marcott, Alan C. Mix, Zhengyu Liu, Bette Otto-Bliesner, Andreas Schmittner,
 and Edouard Bard. 2012. “Global Warming Preceded by Increasing Carbon
 Dioxide Concentrations during the Last Deglaciation.” Nature 484 (7392):
 49–54. doi:10.1038/nature10915.
}

\section{Corrections}{
 None so far required
}
\keyword{datasets}

