% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{sicre.2011.MD99_2275.temperature}
\alias{sicre.2011.MD99_2275.temperature}
\title{Uk'37 based climate proxy for core MD99-2275 published in Sicre et al. (2011)}
\format{
A data frame with 518 rows and 8 variables:
\describe{
  \item{\code{Core}}{}
  \item{\code{Number}}{}
  \item{\code{ID.no}}{}
  \item{\code{ID}}{}
  \item{\code{Proxy.type}}{}
  \item{\code{Published.temperature}}{}
  \item{\code{Age.yrs.AD}}{}
  \item{\code{Age.yrs.BP}}{} 
}
}
\usage{
sicre.2011.MD99_2275.temperature
}
\description{
Uk'37 based climate proxy for core MD99-2275 published in Sicre et al. (2011)
}
\section{Reference}{

 Sicre, M.-A., Hall, I. R., Mignot, J., Khodri, M., Ezat, U., Truong, M.-X., Eiríksson, J., & Knudsen, K.-L. (2011): 
Sea surface temperature variability in the subpolar Atlantic over the last two millennia. Paleoceanography, 26: PA4218.
}

\keyword{datasets}
