% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{climproxyrecords}
\alias{climproxyrecords}
\alias{climproxyrecords-package}
\title{climproxyrecords: A package containing climate proxy data.}
\description{
The climproxyrecords package supplies published proxy climate record data and
compilations. So far it includes data from two compilations:
Marcott et al (2013) and  Shakun et al (2012);

The GISP2 Ice Core Temperature and Accumulation Data from Alley (2000);

Uk'37 based proxy for core MD99-2275 from Sicre et al (2011);

Alkenone based reconstruction for core ODP846 from Herbert et al (2010);

Alkenone based reconstruction for core U1313 from Naafs et al (2012).
}
\details{
For the two compilations the package contains three dataframes: \code{proxies},
\code{dating}, and \code{metadata} (e.g. \code{marcott.proxies}). These
contain the proxy data, associated carbon dating information, and metadata
respectively.

For the individual proxies there are similar dataframes.

\code{alley.accumulation}  

\code{alley.temperature}  


\code{sicre.2011.MD99_2275.age.model} 
 
\code{sicre.2011.MD99_2275.metadata} 
 
\code{sicre.2011.MD99_2275.temperature}  


\code{naafs.U1313.temperature}  

\code{naafs.U1313.metadata}  

\code{herbert.ODP846.temperature}  

\code{herbert.ODP846.metadata}
}
\section{Citing the data}{


For the compilations, please cite both the compilaton papers (Marcott et al 2013, Shakun et al
2012), and the original sources from which the data were extracted. A full
list of sources can be found in the metadata dataframes, e.g. \code{marcott.metadata}.

For the other records please cite the original papers. See the help files e.g.
\code{help(sicre.2011.MD99_2275.temperature)} for the correct citations.
}

