% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{climproxyrecords}
\alias{climproxyrecords}
\alias{climproxyrecords-package}
\title{climproxyrecords: A package containing the climate proxy data from Marcott et
al (2013).}
\description{
The climproxyrecords package supplies proxy climate record data from
published compilations. So far it includes data from two compilations:
Marcott et al (2013) and Shakun et al (2012).
}
\details{
For each compilation the package contains three dataframes: \code{proxies},
\code{dating}, and \code{metadata} (i.e. \code{marcott.proxies}). These
contain the proxy data, associated carbon dating information, and metadata
respectively.

Please cite both the compilaton papers (Marcott et al 2013, Shakun et al
2012), and the original sources from which the data were extracted. A full
list of sources can be found in the dataframes \code{compilation.metadata}.
}

